TAN-93 add governed MCP tool registry metadata
