+++ crates/tandem-server/src/http.rs
@@ -74,6 +74,7 @@ pub(crate) mod governance;
 mod marketplace;
 pub(crate) mod mcp;
 pub(crate) mod mcp_discovery;
+pub(crate) mod mcp_inventory;
 mod middleware;
 mod mission_builder;
 mod mission_builder_host;
