+++ crates/tandem-server/src/http/mcp.rs
@@ -1,14 +1,15 @@
+use super::mcp_inventory::{
+    filter_mcp_inventory_snapshot_to_servers, filter_mcp_snapshot_by_discovery_authorization,
+    filter_mcp_snapshot_by_tool_scope, mcp_inventory_snapshot, session_mcp_tool_filter,
+    McpToolScopeFilter,
+};
 use super::*;
 use base64::Engine;
 use serde::{Deserialize, Serialize};
 use sha2::Digest;
 use tandem_core::tool_name_security_descriptor;
 use tandem_runtime::McpAuthChallenge;
-use tandem_types::{
-    AccessDecision, AccessPermission, DataClass, RequestPrincipal, ResourceKind, ResourceRef,
-    StrictTenantContext, TenantContext, ToolDefaultVisibility, ToolSecurityDescriptor,
-    VerifiedTenantContext,
-};
+use tandem_types::{RequestPrincipal, StrictTenantContext, TenantContext, VerifiedTenantContext};
 use uuid::Uuid;
 
 const BUILTIN_GITHUB_MCP_SERVER_NAME: &str = "github";
@@ -399,113 +400,6 @@ fn strict_context_from_tool_args(args: &Value) -> Option<StrictTenantContext> {
         })
 }
 
-fn mcp_tool_authorized_for_discovery(
-    strict_context: Option<&StrictTenantContext>,
-    tool_name: &str,
-    security: Option<&Value>,
-    now_ms: u64,
-) -> bool {
-    let Some(strict_context) = strict_context else {
-        return true;
-    };
-    let descriptor = security
-        .cloned()
-        .and_then(|value| serde_json::from_value::<ToolSecurityDescriptor>(value).ok())
-        .filter(|descriptor| !descriptor.is_empty())
-        .unwrap_or_else(|| tool_name_security_descriptor(tool_name));
-
-    if descriptor.is_empty() {
-        return true;
-    }
-    if strict_context.is_expired_at(now_ms) {
-        return false;
-    }
-
-    let required_permissions = if descriptor.required_permissions.is_empty() {
-        vec![AccessPermission::View]
-    } else {
-        descriptor.required_permissions.clone()
-    };
-    let data_classes = if descriptor.data_classes.is_empty() {
-        vec![DataClass::Internal]
-    } else {
-        descriptor.data_classes.clone()
-    };
-    let resource_kinds = if descriptor.resource_kinds.is_empty() {
-        vec![ResourceKind::McpTool]
-    } else {
-        descriptor.resource_kinds.clone()
-    };
-
-    let hidden_by_default = matches!(descriptor.default_visibility, ToolDefaultVisibility::Hidden)
-        || descriptor.admin_surface
-        || descriptor.credential_access;
-
-    let all_permissions_allowed = required_permissions.iter().all(|permission| {
-        resource_kinds.iter().any(|resource_kind| {
-            let resource = mcp_tool_resource_ref(strict_context, *resource_kind, tool_name);
-            data_classes.iter().any(|data_class| {
-                matches!(
-                    strict_context
-                        .evaluate_access(&resource, *permission, *data_class, now_ms)
-                        .decision,
-                    AccessDecision::Allow
-                )
-            })
-        })
-    });
-
-    if all_permissions_allowed {
-        return true;
-    }
-
-    !hidden_by_default
-        && required_permissions
-            .iter()
-            .all(|permission| matches!(permission, AccessPermission::View | AccessPermission::Read))
-        && resource_kinds.iter().any(|resource_kind| {
-            let resource = mcp_tool_resource_ref(strict_context, *resource_kind, tool_name);
-            data_classes.iter().any(|data_class| {
-                matches!(
-                    strict_context
-                        .evaluate_access(&resource, AccessPermission::Read, *data_class, now_ms)
-                        .decision,
-                    AccessDecision::Allow
-                ) || matches!(
-                    strict_context
-                        .evaluate_access(&resource, AccessPermission::View, *data_class, now_ms)
-                        .decision,
-                    AccessDecision::Allow
-                )
-            })
-        })
-}
-
-fn mcp_tool_resource_ref(
-    strict_context: &StrictTenantContext,
-    resource_kind: ResourceKind,
-    tool_name: &str,
-) -> ResourceRef {
-    let resource_id = match resource_kind {
-        ResourceKind::McpServer => mcp_server_segment_from_tool_name(tool_name),
-        _ => tool_name.to_string(),
-    };
-    ResourceRef::new(
-        strict_context.tenant_context.org_id.clone(),
-        strict_context.tenant_context.workspace_id.clone(),
-        resource_kind,
-        resource_id,
-    )
-}
-
-fn mcp_server_segment_from_tool_name(tool_name: &str) -> String {
-    let mut parts = tool_name.split('.');
-    match (parts.next(), parts.next()) {
-        (Some("mcp"), Some(server)) if !server.trim().is_empty() => server.to_string(),
-        _ => "mcp".to_string(),
-    }
-}
-
 pub(super) async fn add_mcp(
     State(state): State<AppState>,
     axum::extract::Extension(tenant_context): axum::extract::Extension<TenantContext>,
@@ -580,313 +474,6 @@ fn normalize_mcp_auth_kind(raw: &str) -> String {
         }
         _ => String::new(),
     }
-}
-
-fn mcp_tool_names_for_server(tool_names: &[String], server_name: &str) -> Vec<String> {
-    let prefix = format!("mcp.{}.", mcp_namespace_segment(server_name));
-    let mut tools = tool_names
-        .iter()
-        .filter(|tool_name| tool_name.starts_with(&prefix))
-        .cloned()
-        .collect::<Vec<_>>();
-    tools.sort();
-    tools.dedup();
-    tools
-}
-
-fn catalog_tool_security_by_namespaced_name(
-    catalog_servers: &[Value],
-) -> std::collections::HashMap<String, Value> {
-    let mut out = std::collections::HashMap::new();
-    for server in catalog_servers {
-        let Some(tool_security) = server.get("tool_security").and_then(Value::as_object) else {
-            continue;
-        };
-        for value in tool_security.values() {
-            let Some(namespaced_name) = value.get("namespaced_name").and_then(Value::as_str) else {
-                continue;
-            };
-            if let Some(security) = value.get("security") {
-                out.insert(namespaced_name.to_string(), security.clone());
-            }
-        }
-    }
-    out
-}
-
-fn mcp_tool_security_for_inventory_server(
-    server_name: &str,
-    remote_tool_names: &[String],
-    registered_tool_names: &[String],
-    catalog_tool_security: &std::collections::HashMap<String, Value>,
-    registered_security: &std::collections::HashMap<String, Value>,
-) -> Value {
-    let server_segment = mcp_namespace_segment(server_name);
-    let mut tool_names = remote_tool_names
-        .iter()
-        .chain(registered_tool_names.iter())
-        .cloned()
-        .collect::<Vec<_>>();
-    tool_names.sort();
-    tool_names.dedup();
-
-    let mut out = serde_json::Map::new();
-    for tool_name in tool_names {
-        let short_name = tool_name
-            .strip_prefix(&format!("mcp.{server_segment}."))
-            .unwrap_or(&tool_name)
-            .to_string();
-        let security = catalog_tool_security
-            .get(&tool_name)
-            .cloned()
-            .filter(security_value_non_empty)
-            .or_else(|| {
-                registered_security
-                    .get(&tool_name)
-                    .cloned()
-                    .filter(security_value_non_empty)
-            })
-            .unwrap_or_else(|| {
-                serde_json::to_value(tool_name_security_descriptor(&tool_name))
-                    .unwrap_or(Value::Null)
-            });
-        out.insert(
-            short_name.clone(),
-            json!({
-                "tool_name": short_name,
-                "namespaced_name": tool_name,
-                "security": security,
-            }),
-        );
-    }
-    Value::Object(out)
-}
-
-fn security_value_non_empty(value: &Value) -> bool {
-    serde_json::from_value::<ToolSecurityDescriptor>(value.clone())
-        .map(|descriptor| !descriptor.is_empty())
-        .unwrap_or(false)
-}
-
-#[derive(Default)]
-struct McpToolScopeFilter {
-    wildcard_server_segments: std::collections::HashSet<String>,
-    exact_tool_names: std::collections::HashSet<String>,
-}
-
-fn parse_mcp_tool_scope_filter(tool_names: &[String]) -> McpToolScopeFilter {
-    let mut filter = McpToolScopeFilter::default();
-    for raw in tool_names {
-        let tool_name = raw.trim();
-        if tool_name.is_empty() {
-            continue;
-        }
-        if let Some(rest) = tool_name.strip_prefix("mcp.") {
-            if let Some((server_segment, tool_segment)) = rest.split_once('.') {
-                if tool_segment == "*" {
-                    filter
-                        .wildcard_server_segments
-                        .insert(server_segment.to_string());
-                } else {
-                    filter
-                        .exact_tool_names
-                        .insert(format!("mcp.{server_segment}.{tool_segment}"));
-                }
-            }
-        }
-    }
-    filter
-}
-
-fn filter_mcp_snapshot_by_tool_scope(snapshot: Value, filter: &McpToolScopeFilter) -> Value {
-    let mut snapshot = snapshot;
-    if filter.wildcard_server_segments.is_empty() && filter.exact_tool_names.is_empty() {
-        return snapshot;
-    }
-
-    if let Some(root) = snapshot.as_object_mut() {
-        if let Some(Value::Array(rows)) = root.get_mut("servers") {
-            rows.retain(|row| {
-                let server_name = row.get("name").and_then(Value::as_str).unwrap_or("");
-                let server_segment = mcp_namespace_segment(server_name);
-                if filter.wildcard_server_segments.contains(&server_segment) {
-                    return true;
-                }
-                let exact_tools = row
-                    .get("remote_tools")
-                    .and_then(Value::as_array)
-                    .map(|tools| {
-                        tools
-                            .iter()
-                            .filter_map(Value::as_str)
-                            .map(str::to_string)
-                            .collect::<std::collections::HashSet<_>>()
-                    })
-                    .unwrap_or_default();
-                exact_tools
-                    .iter()
-                    .any(|tool_name| filter.exact_tool_names.contains(tool_name))
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("connected_server_names") {
-            rows.retain(|row| {
-                row.as_str().is_some_and(|server| {
-                    let segment = mcp_namespace_segment(server);
-                    filter.wildcard_server_segments.contains(&segment)
-                        || filter
-                            .exact_tool_names
-                            .iter()
-                            .any(|tool| tool.starts_with(&format!("mcp.{segment}.")))
-                })
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("enabled_server_names") {
-            rows.retain(|row| {
-                row.as_str().is_some_and(|server| {
-                    let segment = mcp_namespace_segment(server);
-                    filter.wildcard_server_segments.contains(&segment)
-                        || filter
-                            .exact_tool_names
-                            .iter()
-                            .any(|tool| tool.starts_with(&format!("mcp.{segment}.")))
-                })
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("remote_tools") {
-            rows.retain(|row| {
-                row.get("server_name")
-                    .and_then(Value::as_str)
-                    .is_some_and(|server| {
-                        let segment = mcp_namespace_segment(server);
-                        if filter.wildcard_server_segments.contains(&segment) {
-                            return true;
-                        }
-                        row.get("namespaced_name")
-                            .and_then(Value::as_str)
-                            .is_some_and(|tool_name| filter.exact_tool_names.contains(tool_name))
-                    })
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("registered_tools") {
-            rows.retain(|row| {
-                row.as_str().is_some_and(|tool_name| {
-                    tool_name == "mcp_list"
-                        || filter.exact_tool_names.contains(tool_name)
-                        || filter
-                            .wildcard_server_segments
-                            .iter()
-                            .any(|segment| tool_name.starts_with(&format!("mcp.{segment}.")))
-                })
-            });
-        }
-    }
-
-    snapshot
-}
-
-pub(crate) async fn mcp_inventory_snapshot(state: &AppState) -> Value {
-    let mut server_rows = state.mcp.list().await.into_values().collect::<Vec<_>>();
-    server_rows.sort_by(|a, b| a.name.cmp(&b.name));
-
-    let remote_tools = state.mcp.list_tools().await;
-    let registered_schemas = state.tools.list().await;
-    let registered_tool_names = registered_schemas
-        .iter()
-        .map(|schema| schema.name.clone())
-        .collect::<Vec<_>>();
-    let registered_security_by_name = registered_schemas
-        .into_iter()
-        .map(|schema| {
-            (
-                schema.name.clone(),
-                serde_json::to_value(schema.security).unwrap_or(Value::Null),
-            )
-        })
-        .collect::<std::collections::HashMap<_, _>>();
-    let catalog_tool_security = crate::mcp_catalog::index()
-        .and_then(|catalog| catalog.get("servers"))
-        .and_then(Value::as_array)
-        .map(|servers| catalog_tool_security_by_namespaced_name(servers.as_slice()))
-        .unwrap_or_default();
-
-    let mut connected_server_names = Vec::new();
-    let mut enabled_server_names = Vec::new();
-    let mut all_remote_tool_names = Vec::new();
-    let mut all_registered_tool_names = Vec::new();
-    let mut servers = Vec::new();
-
-    for server in server_rows {
-        let mut remote_tool_names = remote_tools
-            .iter()
-            .filter(|tool| tool.server_name == server.name)
-            .map(|tool| tool.namespaced_name.trim().to_string())
-            .filter(|tool_name| !tool_name.is_empty())
-            .collect::<Vec<_>>();
-        remote_tool_names.sort();
-        remote_tool_names.dedup();
-
-        let registered_names = mcp_tool_names_for_server(&registered_tool_names, &server.name);
-        let tool_security = mcp_tool_security_for_inventory_server(
-            &server.name,
-            &remote_tool_names,
-            &registered_names,
-            &catalog_tool_security,
-            &registered_security_by_name,
-        );
-
-        if server.enabled {
-            enabled_server_names.push(server.name.clone());
-        }
-        if server.connected {
-            connected_server_names.push(server.name.clone());
-        }
-        all_remote_tool_names.extend(remote_tool_names.clone());
-        all_registered_tool_names.extend(registered_names.clone());
-
-        let mut pending_auth_tools = server
-            .pending_auth_by_tool
-            .keys()
-            .cloned()
-            .collect::<Vec<_>>();
-        pending_auth_tools.sort();
-        pending_auth_tools.dedup();
-
-        servers.push(json!({
-            "name": server.name,
-            "transport": server.transport,
-            "enabled": server.enabled,
-            "connected": server.connected,
-            "last_error": server.last_error,
-            "last_auth_challenge": server.last_auth_challenge,
-            "pending_auth_tools": pending_auth_tools,
-            "remote_tool_count": remote_tool_names.len(),
-            "registered_tool_count": registered_names.len(),
-            "allowed_tool_count": server.allowed_tools.as_ref().map(|tools| tools.len()).unwrap_or(remote_tool_names.len()),
-            "allowed_tools": server.allowed_tools.clone(),
-            "discovered_tool_count": server.tool_cache.len(),
-            "remote_tools": remote_tool_names,
-            "registered_tools": registered_names,
-            "tool_security": tool_security,
-        }));
-    }
-
-    connected_server_names.sort();
-    connected_server_names.dedup();
-    enabled_server_names.sort();
-    enabled_server_names.dedup();
-    all_remote_tool_names.sort();
-    all_remote_tool_names.dedup();
-    all_registered_tool_names.sort();
-    all_registered_tool_names.dedup();
-
-    json!({
-        "inventory_version": 1,
-        "connected_server_names": connected_server_names,
-        "enabled_server_names": enabled_server_names,
-        "remote_tools": all_remote_tool_names,
-        "registered_tools": all_registered_tool_names,
-        "servers": servers,
-    })
 }
 
 async fn current_mcp_auth_challenge(state: &AppState, name: &str) -> Option<McpAuthChallenge> {
@@ -1524,231 +1111,6 @@ async fn finish_mcp_oauth_callback(
         session_mut.error = None;
     }
     Ok(())
-}
-
-fn filter_mcp_inventory_snapshot_to_servers(snapshot: Value, allowed_servers: &[String]) -> Value {
-    let mut snapshot = snapshot;
-    let allowed_servers = allowed_servers
-        .iter()
-        .map(|server| server.trim().to_string())
-        .filter(|server| !server.is_empty())
-        .collect::<std::collections::HashSet<_>>();
-    if allowed_servers.is_empty() {
-        return snapshot;
-    }
-    let allowed_tool_prefixes = allowed_servers
-        .iter()
-        .map(|server| format!("mcp.{}.", mcp_namespace_segment(server)))
-        .collect::<Vec<_>>();
-
-    let keep_server = |name: &str| allowed_servers.contains(name);
-
-    if let Some(root) = snapshot.as_object_mut() {
-        if let Some(Value::Array(rows)) = root.get_mut("servers") {
-            rows.retain(|row| {
-                row.get("name")
-                    .and_then(Value::as_str)
-                    .is_some_and(keep_server)
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("connected_server_names") {
-            rows.retain(|row| row.as_str().is_some_and(keep_server));
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("enabled_server_names") {
-            rows.retain(|row| row.as_str().is_some_and(keep_server));
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("remote_tools") {
-            rows.retain(|row| {
-                row.get("server_name")
-                    .and_then(Value::as_str)
-                    .is_some_and(keep_server)
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("registered_tools") {
-            rows.retain(|row| {
-                row.as_str().is_some_and(|tool_name| {
-                    tool_name == "mcp_list"
-                        || allowed_tool_prefixes
-                            .iter()
-                            .any(|prefix| tool_name.starts_with(prefix))
-                })
-            });
-        }
-    }
-
-    snapshot
-}
-
-fn filter_mcp_snapshot_by_exact_and_wildcard_tools(
-    snapshot: Value,
-    allowed_tools: &[String],
-) -> Value {
-    let filter = parse_mcp_tool_scope_filter(allowed_tools);
-    filter_mcp_snapshot_by_tool_scope(snapshot, &filter)
-}
-
-/// Filter MCP inventory by namespace segments (e.g. `["tandem_mcp"]`) derived
-/// from `session_allowed_tools` patterns like `mcp.tandem_mcp.*`.  Server names
-/// are matched by applying `mcp_namespace_segment` so that `"tandem-mcp"` matches
-/// the segment `"tandem_mcp"`.
-fn filter_mcp_snapshot_by_namespace_segments(
-    snapshot: Value,
-    allowed_segments: &[String],
-) -> Value {
-    let mut snapshot = snapshot;
-    let segments_set: std::collections::HashSet<&str> =
-        allowed_segments.iter().map(|s| s.as_str()).collect();
-    let keep_server = |name: &str| segments_set.contains(mcp_namespace_segment(name).as_str());
-    let allowed_tool_prefixes: Vec<String> = allowed_segments
-        .iter()
-        .map(|seg| format!("mcp.{}.", seg))
-        .collect();
-
-    if let Some(root) = snapshot.as_object_mut() {
-        if let Some(Value::Array(rows)) = root.get_mut("servers") {
-            rows.retain(|row| {
-                row.get("name")
-                    .and_then(Value::as_str)
-                    .is_some_and(keep_server)
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("connected_server_names") {
-            rows.retain(|row| row.as_str().is_some_and(keep_server));
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("enabled_server_names") {
-            rows.retain(|row| row.as_str().is_some_and(keep_server));
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("remote_tools") {
-            rows.retain(|row| {
-                row.get("server_name")
-                    .and_then(Value::as_str)
-                    .is_some_and(keep_server)
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("registered_tools") {
-            rows.retain(|row| {
-                row.as_str().is_some_and(|tool_name| {
-                    tool_name == "mcp_list"
-                        || allowed_tool_prefixes
-                            .iter()
-                            .any(|prefix| tool_name.starts_with(prefix))
-                })
-            });
-        }
-    }
-    snapshot
-}
-
-fn filter_mcp_snapshot_by_discovery_authorization(
-    snapshot: Value,
-    strict_context: Option<&StrictTenantContext>,
-    now_ms: u64,
-) -> Value {
-    if strict_context.is_none() {
-        return snapshot;
-    }
-    let mut snapshot = snapshot;
-    if let Some(root) = snapshot.as_object_mut() {
-        let mut allowed_tools = std::collections::HashSet::<String>::new();
-        let mut allowed_server_segments = std::collections::HashSet::<String>::new();
-
-        if let Some(Value::Array(rows)) = root.get_mut("servers") {
-            rows.retain_mut(|row| {
-                let server_name = row.get("name").and_then(Value::as_str).unwrap_or("");
-                let server_segment = mcp_namespace_segment(server_name);
-                let tool_security = row
-                    .get("tool_security")
-                    .and_then(Value::as_object)
-                    .cloned()
-                    .unwrap_or_default();
-
-                let mut server_allowed_tools = std::collections::HashSet::<String>::new();
-                for field in ["remote_tools", "registered_tools"] {
-                    if let Some(Value::Array(tools)) = row.get_mut(field) {
-                        tools.retain(|tool| {
-                            let Some(tool_name) = tool_name_from_inventory_value(tool) else {
-                                return false;
-                            };
-                            if tool_name == "mcp_list" {
-                                return true;
-                            }
-                            let short_name = tool_name
-                                .strip_prefix(&format!("mcp.{server_segment}."))
-                                .unwrap_or(&tool_name);
-                            let security = tool_security
-                                .get(short_name)
-                                .and_then(|value| value.get("security"));
-                            let allowed = mcp_tool_authorized_for_discovery(
-                                strict_context,
-                                &tool_name,
-                                security,
-                                now_ms,
-                            );
-                            if allowed {
-                                allowed_tools.insert(tool_name.clone());
-                                server_allowed_tools.insert(tool_name);
-                            }
-                            allowed
-                        });
-                    }
-                }
-
-                let keep = !server_allowed_tools.is_empty();
-                if keep {
-                    allowed_server_segments.insert(server_segment);
-                }
-                keep
-            });
-        }
-
-        if let Some(Value::Array(rows)) = root.get_mut("remote_tools") {
-            rows.retain(|row| {
-                tool_name_from_inventory_value(row)
-                    .is_some_and(|tool_name| allowed_tools.contains(&tool_name))
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("registered_tools") {
-            rows.retain(|row| {
-                row.as_str().is_some_and(|tool_name| {
-                    tool_name == "mcp_list" || allowed_tools.contains(tool_name)
-                })
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("connected_server_names") {
-            rows.retain(|row| {
-                row.as_str().is_some_and(|server| {
-                    allowed_server_segments.contains(&mcp_namespace_segment(server))
-                })
-            });
-        }
-        if let Some(Value::Array(rows)) = root.get_mut("enabled_server_names") {
-            rows.retain(|row| {
-                row.as_str().is_some_and(|server| {
-                    allowed_server_segments.contains(&mcp_namespace_segment(server))
-                })
-            });
-        }
-    }
-    snapshot
-}
-
-fn tool_name_from_inventory_value(value: &Value) -> Option<String> {
-    value
-        .as_str()
-        .map(str::to_string)
-        .or_else(|| {
-            value
-                .get("namespaced_name")
-                .or_else(|| value.get("name"))
-                .and_then(Value::as_str)
-                .map(str::to_string)
-        })
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-}
-
-fn session_mcp_tool_filter(session_tools: &[String]) -> McpToolScopeFilter {
-    parse_mcp_tool_scope_filter(session_tools)
 }
 
 async fn scoped_mcp_servers_for_session(state: &AppState, session_id: &str) -> Vec<String> {
